requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
dataclasses-json==0.5.14
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
faiss-cpu==1.7.4
fastapi==0.95.2
filelock==3.18.0
flask==2.3.2
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
gitpython==3.1.44
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.0.238
langsmith==0.0.92
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
narwhals==1.40.0
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.2.3
pdfminer-six==20221105
pdfplumber==0.9.0
pillow==9.5.0
propcache==0.3.1
protobuf==4.25.7
pyarrow==20.0.0
pycparser==2.22
pydantic==1.10.22
pydeck==0.9.1
pygments==2.19.1
pymongo==4.3.3
pympler==1.1
pypdf==5.5.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rpds-py==0.25.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.27.0
streamlit==1.24.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.5
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspect==0.9.0
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.4.0
uvicorn==0.22.0
validators==0.35.0
wand==0.6.13
werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
